@@ -3,12 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.5.1",
+    "@fortawesome/free-brands-svg-icons": "^6.5.1",
+    "@fortawesome/free-regular-svg-icons": "^6.5.1",
+    "@fortawesome/free-solid-svg-icons": "^6.5.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "react": "^18.3.0-canary-98b8359f6-20240223",
-    "react-dom": "^18.3.0-canary-98b8359f6-20240223",
-    "react-router-dom": "^6.22.1",
+    "@types/react": "^18.2.61",
+    "@types/react-dom": "^18.2.19",
+    "postcss-import": "^16.0.1",
+    "postcss-nested": "^6.0.1",
+    "react": "^18.3.0-canary-2f8f77602-20240229",
+    "react-dom": "^18.3.0-canary-2f8f77602-20240229",
+    "react-router-dom": "^6.22.3",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
@@ -18,7 +27,9 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "lint": "eslint 'src.{js,ts,tsx}'",
-    "lint:fix": "eslint --fix 'src.{js,ts,tsx}'"
+    "lint:fix": "eslint --fix 'src.{js,ts,tsx}'",
+    "format": "prettier --write \"src/**/*.tsx\"",
+    "build": "webpack --mode production"
   },
   "eslintConfig": {
     "extends": [
@@ -40,13 +51,23 @@
   },
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@types/tailwindcss": "^3.1.0",
     "@typescript-eslint/eslint-plugin": "^7.0.2",
     "@typescript-eslint/parser": "^7.0.2",
+    "autoprefixer": "^10.4.18",
+    "css-loader": "^6.10.0",
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
+    "file-loader": "^6.2.0",
     "or": "^0.2.0",
+    "postcss": "^8.4.35",
     "prettier": "^3.2.5",
-    "typescript": "^5.3.3"
+    "style-loader": "^3.3.4",
+    "tailwindcss": "^3.4.1",
+    "ts-loader": "^9.5.1",
+    "typescript": "^5.4.2",
+    "webpack": "^5.90.3",
+    "webpack-cli": "^5.1.4"
   }
 }
