@@ -1,18 +1,19 @@
+// src/index.tsx
 import React from 'react';
-import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 import App from './App';
+import reportWebVitals from './reportWebVitals';
+import { createRoot } from 'react-dom/client'; // Import createRoot
 
-const root = document.getElementById('root');
+const root = document.getElementById('root') as HTMLElement;
 if (root !== null) {
-  // @ts-ignore
-  // eslint-disable-next-line react/no-deprecated
-  ReactDOM.render(
+  createRoot(root).render(
     <React.StrictMode>
       <Router>
         <App />
       </Router>
     </React.StrictMode>,
-    root,
   );
 }
+
+reportWebVitals;
