@@ -1,17 +1,14 @@
+// main.tsx
 import React from 'react';
-import './Main.css';
+import HeroSection from './HeroSection';
+import SectionA from './sections/SectionA';
 
-interface MainProps {
-  className?: string;
-}
-
-const Main: React.FC<MainProps> = ({ className }) => {
+const Main: React.FC = () => {
   return (
-    <main className={className}>
-      <h1>Welcome to Robby's Portfolio</h1>
-      <p>This is a simple portfolio website built using React and Express.</p>
-      <p>Feel free to explore the projects and get in touch!</p>
-    </main>
+    <div>
+      <HeroSection />
+      <SectionA />
+    </div>
   );
 };
 
