@@ -1,20 +1,27 @@
-// Path: frontend/robby/src/components/NavBar.css
+// src/components/Navbar.tsx
+
 import React from 'react';
-import './NavBar.css';
 import { Link } from 'react-router-dom';
+import './NavBar.css';
 
 const Navbar = () => {
   return (
-    <nav className="navbar">
-      <ul className="nav-links">
+    <nav>
+      <ul className="navbar-list">
         <li>
-          <Link to="/doc">Documentation</Link>
+          <Link to="/">Home</Link>
         </li>
         <li>
-          <Link to="/blog">Blog</Link>
+          <Link to="/community">Community</Link>
         </li>
         <li>
           <Link to="/about">About</Link>
+        </li>
+        <li>
+          <Link to="/contact">Contact</Link>
+        </li>
+        <li>
+          <Link to="/blog">Blog</Link>
         </li>
       </ul>
     </nav>
