@@ -1,10 +1,15 @@
+// AboutPage.tsx
 import React from 'react';
+import './AboutPage.css'; // Import the associated CSS file
 
 const AboutPage: React.FC = () => {
   return (
-    <div>
-      <h1>About Us</h1>
-      <p>This is the about page of our application.</p>
+    <div className="about-page-container">
+      <h2>About Us</h2>
+      <p>
+        This is the about page. You can use it to describe your company, team,
+        or product.
+      </p>
     </div>
   );
 };
